Avoid segfault when meta-parsing a null program.
The result of parsing the null program has always properly been I, the identity
function.  Formerly, it was returning the static instance of "I" declared
inside parse.c -- the same one that gets dropped after the parse is finished.
Normally that's not a problem because with a non-null program that "I" is
embedded inside a larger structure, so its reference count remains > 0 after
the final drop.  But with a null program, the reference count would drop
to 0, the value would get recycled, and *then* we'd attempt to evaluate it.
So now the code returns a fresh instance of "I", created with Q(fexl_I).

Note that I only observed this problem while developing test cases for meta-
parsing in combination with the "show" routine written in Fexl.  Only then did
it arise.  Before that, I was always able to parse the null program just fine.
Must have something to do with needing a new value structure at just the
"right" time to trigger the issue.

That's a long commit message for a three-line code change, but it bears
mentioning because I'm always striving to make Fexl rock-solid and avoid all
such defects.