Refactor for compactness and elegance; now about 25% faster too.
The value.c file is very clean, with D, Q, and A all defined in terms of a
static "create" function.  This also streamlined the implementation of the
data types long, double, and string.

I added a built in "append" function, which I did by calling the Fexl parser
on a piece of static text in the C program.

Minimized dependency on stdio.h even further.

Cleaned up comments in basic.c.