Update a1.fxl to use the "continuation style" throughout.
Here's an example of the continuation style:

	\hello =
	(\next
	print "Hello world."; nl;
	next
	)

You can do things like this, with the result you'd expect:

	hello; hello; hello;

However, if you didn't chain the hello function with the "next" parameter, it
would not work properly because the print would only happen *once*, like a
"one-shot" function.  Consider:

	\hello ==
	(
	print "Hello world."; nl;
	)

	hello; hello; hello;

When the first hello is evaluated, it has the desired side effect, but then
hello is *replaced* with its final value, namely the identity function.
Consequently, subsequent calls to hello have no effect.

When you chain calls with the next paramter, in effect you create a fresh copy
of the function which can be reduced in place, with side effects, each time you
call it.  Note that the "copy" is very efficient, only replacing the value of
the next parameter but sharing everything else, because it uses the lambda
substitution code that shares sub-trees.