No need to change the stack technique in parse.c (i.e., no performance boost).
Low-level testing confirms that the ability to rely on the assumption that
f->R != 0 whenever f->L != 0 does not increase the speed of the clear routine
in value.c.  In fact, during testing I discovered strangely that this code
without the extra check:

	if (--f->R->N == 0)
		recycle(f->R);

Was actually *slower* than this code with the check:

	if (f->R && --f->R->N == 0)
		recycle(f->R);

Go figure.  I guess with modern optimizers, caches, and branch predictors you
can see some surprising things.