Fixed two segmentation fault cases.
I fixed the "resolve" routine so it doesn't fault when given a sym that's not a
name or string.

I fixed the "main" routine so it doesn't fault when you run Fexl on a file that
doesn't exist.  I unwittingly let that case drop several revisions ago when I
cleaned up the parse and resolve code.  Ouch!  That did give me an excuse to
clean up main and create a separate "parse_main" to handle separately the
parse, resolve, and error reporting.