Fexl now uses /usr/share/fexl/top.fxl as the standard context for all scripts. This gives Fexl a potentially large library of built-in functions written in Fexl itself, in addition to the functions written in C (in the form of routines named fexl_X, where X is the function name).
The directory structure for the installed fexl is now:

  /usr/bin/fexl
  /usr/lib/fexl/
  /usr/share/fexl/

The libfexl.so library is now inside /usr/lib/fexl instead of /usr/lib.

The /usr/share/fexl directory contains top.fxl.  This directory is overwritten
completely when you do an install, so you should not edit its contents directly.
If you like, you can edit top.fxl in your local build directory, and that will
be copied up when you do the install.  In effect you would be forking your own
standard context, no different in principle from editing the C code itself.

The top.fxl file provides a "sky's the limit" bootstrap capability, so now it's
just a matter of rapidly expanding the library.  We can even do advanced things
like fetch definitions from the network, or install packages with apt-get.

This is how we do "modules": by parsing Fexl text inside a context which
provides definitions for all the undefined symbols referenced in the text.
This is also how we provide "safe" or "restricted" environments -- for example
in a web context you might want "print" not to go to stdout, but into a string
instead, and you wouldn't want "delete_file" to be defined (at least not in its
normal sense).

Because we're now using top.fxl, I was able to remove a lot of extra function
definitions from the test scripts test/a1.fxl and test/a2.fxl.  I also removed
the fexl_append function from basic.c, because now append is simply defined in
top.fxl.

I also dramatically simplified the error reporting in die.c.  I was kind of
backed into a corner at one point, because I was finding it difficult to report
the correct script name when an error occurred, particularly if the error
occurred in top.fxl itself, because then the script name wasn't even in the
argv array!  The way I do it now always reports the script name perfectly, and
the warn routine in die.c is now very generic and simple, and doesn't have to
know anything about Fexl itself.  Note in particular that it no longer refers
to a global "line" variable to report a line number.  I also don't need "argi"
anymore, so I took that out of run.c and meta.c.

Both the "line" and "error" variables are now static within parse.c, with no
need for them to be global now.

I also don't need the canonical_name routine in resolve.c anymore, since those
synonyms are now provided in top.fxl instead.