Standard context now defined in terms of elegant combinators.
In share/fexl, starting with main.fxl, the standard context is now defined with
neat combinators "define", "starts", "void", and "see".

This set of combinators makes it very easy to build indexes which match on a
prefix of the symbol being resolved.  This makes the standard context easy to
scale efficiently even with many thousands of built-in symbols, because we can
partition the name space on the basis of common prefixes to as much depth as
we need to make the search efficient.

Also, formerly I had defined "resolve" to refer to the enhanced context defined
in main.fxl, instead of the inner context defined in resolve.c.  I change that
back.  Now "resolve" refers to the resolve.c context, and the new name
"standard" refers to the enhanced context defined in main.fxl.  This simplifies
the definition of "standard" in main.fxl, and gives you more flexibility.