Start splitting the name space alphabetically.
I've added index files like "a.fxl", "m.fxl", etc.  It's a bit overkill at this
point, since the standard name space is still fairly small.  But it's good to
go ahead and test it, and it's working really well.

I also added a test for "rand" in a1.fxl.