Because the "resolve_parse" function defined in test/a1.fxl was so similar to the standard version used in share/fexl/main.fxl, I forcibly unified the two routines so now a1.fxl uses the one defined in main.fxl.  I don't like having "resolve_parse" as a standard function because I don't like its interface or abstraction level, so this is a temporary measure until I can clean up.
This is a precursor for possibly eliminating the data structure traversal used
in src/fexl.c to walk the parse result and report any errors.  I may be able to
hook right into the Fexl code in main.fxl which already does that, even though
at first glance it seems like a chicken-and-egg problem.