Fixed problem in parse.c line 442 where I drop def, but def happens to be equal to val because we hit the special I pattern in the call to apply above it.  In that case we *need* val and don't want to discard it.  Notice that this fixes two test cases in a1.fxl which I stupidly didn't check carefully when I did the previous commit that removed implicit eager semantics from definitions.
I despise obscure issues like this, preferring self-evident clarity in the code
everywhere.  The good news is, I'm currently working on a separate branch where
all my parse functions return a value of the form (pair ok; pair exp; symbols).
The purpose of symbols is to compute the lambda substitution patterns from the
bottom up, instead of reaching down from the top in get_pattern.

A side effect of this technique is that we should no longer have obscure cases
due to optimizations, since we'll always be holding references inside the new
higher-level values being returned from the parse functions, instead of, as in
the case of this recent bug, inadvertently dropping a reference to something we
still need.