To avoid the despised obscure issues I mentioned in the last commit, I re- factored the parse code so everything looks a lot more natural and compact, in particular parse_lambda.  There aren't as many arbitrary looking hold and drop operations, because now the "apply" and "lambda" routines are smart about doing the hold and drop on their arguments.  Also, because they handle NULL cases intelligently, the various parse results can be chained together more simply without so many conditional checks, even when there's a syntax error.
You'll also notice that the "push" and "pop" calls are more isolated.  In an
upcoming version we won't need any of that because we'll be computing the
lambda patterns from the bottom-up instead of top-down as we do now in
get_pattern.

Furthermore, the lambda routine now calls a new "copy_exp" routine to make a
shallow copy of f->L, instead of wrapping it inside A(I,f->L) instead.  Note
that we have to do *something* special because this is a case where the result
f->L is a subcomponent of the parent structure f, and we're about to *drop* f
because it's no longer used.  That would be a dangerous operation unless we
handle it consciously and correctly.  The nice thing about doing the copy
instead of applying I to it is that it greatly compresses some of the parsed
forms, as you can see by looking at the diffs in a1.out.  Unfortunately a1.out
looks like a binary file in github because it has some control characters, so
it's hard to see the diffs on github.  But if you look into it, you can see
some serious reductions.

For example, this form:
	(((((lam (C (C (I C)))) ((lam (C (C I))) (I (long_add 4)))) long_add) 4) 5)

is now simply (((I long_add) 4) 5).

Also, this form:
	((lam (C (C (I C)))) ((lam (C (C I))) (I (<= ==))))

is now simply I.

You may also notice that in value.c I've eliminated the "create" routine.
Formerly the A, D, and Q routines were defined in terms of create.  Strangely,
when I added the new copy routine, and defined it in terms of create, I found
that the a2 test slowed down, taking 19.5 seconds to run instead of 16.8!  Even
more strangely, it was the mere *presence* of the new copy routine which caused
the slowdown -- I didn't even have to call it or use it in any way!  Evidently
because copy was calling create, it pushed something over the edge in the C
compiler so that create was no longer being inlined, and we suddenly saw extra
function call overhead.

To work around that issue, I flipped things on its head so now "A" is the
primary routine, and D, Q, and copy are defined in terms of A.