After installing as sudo, set permissions on local bin and obj to $SUDO_USER.
So, if you're installing under sudo, this changes the owner and group on the
local bin and obj directories to the non-privileged "true" user $SUDO_USER.
Otherwise you'd have files owned by root in your local directory, which is
annoying.

A clear-cut test case which illustrates the advantage of this change is to do
the following two commands in immediate succession:

	sudo ./build quiet clean install
	./build quiet clean install in ~/myfexl

Before the change, you'd get a bunch of errors on the second command as it
tried to delete object files owned by root created in the first command.  With
the change, everything works smoothly.