You can now call readlink(2) from Fexl, for example:
	readlink "/proc/self/exe" \path
	print "You are currently running ";print path;nl;

I also changed the build script slightly so that its dependency analysis (via
grep) does not insist that #include lines always end precisely in a double-
quote.  For example, you can now add spaces or comments on the end of #include
lines, and the grep ignores them.  I had just made it too strict before, by
including '$' at the end of the pattern to match end-of-line immediately after
the quote.