Split more functions into libraries so they're not linked into bin/fexl. The executable is now about 27K.
Included more test cases in test/a1.fxl.

Updated the build script to allow libraries to be linked with other libraries.
For example, see "lib qstr2 - str2" in the config file.  I also enhanced
make_bin to understand that it doesn't need to re-link the executable just
because a dynamic library changed.