Preparing Fexl for full modularity.
The enhanced context (fexl.fxl) now defines a symbol "standard_context", which
consists of the core_context and the enhanced_context chained together.  The
core_context is the names defined in the C code, and the enhanced_context is
the names defined in fexl.fxl itself.  Together these constitute the
standard_context, which is the full environment available to the top-level
Fexl program.

The test/a1.fxl program now starts by parsing and resolving the remaining code
after the first occurrence of "\\".  At the moment this is just "showing off",
i.e. there's no particular purpose for doing it.  However, it illustrates an
idiom which I will use to implement full modularity in Fexl, in the form of a
"use" function.  The "use" function will allow you to use external definitions
as the context for the code which follows "\\".

I also made a few cosmetic changes in resolve.c, as follows.
Renamed "define_name" as "core_context".
Renamed "resolve_source" as "source_context".