The "use" function is now easy to use.
The code is still not complete, and it's not as elegant and serene as I would
like, but this is progress.  Fexl is basically the "Plan 9" of programming
languages, and surprisingly the most difficult part is reporting the correct
source name and line number in error messages!  I think the code will clean up
pretty nicely though.