Defined setrlimit and the RLIMIT constants in Fexl.
Eliminated fexl.h and moved those declarations to system.h.  We don't want to
have a .h file which corresponds to a main .c program because if you include
that .h file from a different main program, the build system would link in the
.o and give you multiple definitions of "main".

Enhanced the build script to notice when a cache/o entry is older than any of
the .o files it contains.