Made Fexl highly portable and lean (under 32K now).
Added full-fledged arithmetic and comparison operators.

New values are created with reference count 1 instead of 0, which gives the
code some reassuring properties when returning a value that is under another
value.

Removed the free list.  Unused values are completely freed now.

Eliminated the use of printf and stdarg.

The new approach in Fexl is *not* to make the raw machine details visible to
the Fexl programmer, but rather to insulate the programmer from such details.
"Don't master complexity, avoid it."