Make lambda definition do eager evaluation by default.
The defines y as the final value of (f x), which is evaluated first:

	\y=(f x)

This defines y as the form (f x), which is evaluated in every place y appears
subsequently:

	\y=\(f x)

Of course, if the definition takes a parameter, you can still simply say:

	\y=(\x f x)

In making this change, I was able to eliminate type_later entirely, which in
turn greatly simplified type_query.  It also reduced the run time of some of my
Fexl programs by about 30%, evidently because I was not specifying eager
evaluation as often as I could have.  Now that eager evaluation is the default,
I don't have to think about that.

One more minor thing:  I renamed Qcons as cons.