Disable core dump when running badly behaved programs.
When I ran certain badly behaved programs like "@ @" on some older machines,
it would produce a core dump file when it died.  Using "ulimit -Sc 0" prevents
that.

I also trimmed back the setting on ulimit -Sv so that, when running on a
slower machine, the program will run out of memory before it runs out of time.