Improve the "get" function.
I did a couple of things there.  First, the old version was doing a replace_A.
That's not wise when doing functions with side effects.  In some new test code,
not included here, I could devise cases where the wrong thing happened.  In
particular if I started with one input stream and then switched to another,
the first character of the first input stream could be returned again after
the switch.  Eliminating the replace_A avoided that.

Second, I simplified type_get so it just returns {ch}, instead of requiring a
handler argument to which it would pass the result.

I also made some cosmetic changes to standard.c.