Remove eager evaluation from lambda assignment.
Now a lambda assignment is just a simple symbol definition.

This eliminates the need for the "query" and "later" functions.  It also
requires that functions which return a different value every time, such as
"rand" and "get", must now return a tuple, so you can do this:

	rand \x   # x is the next random number
	get \ch   # ch is the next character from the input

Before, when lambda assignment did eager evaluation, we could do this:

	\x=rand   # x is the next random number
	\ch=get   # ch is the next character from the input

Now all that does is define "x" as a synonym for "rand", and "ch" as a synonym
for "get" -- which is almost certainly NOT what we want.

This change simplifies the language semantics, and simplifies the working Fexl
code I have in production.  It is conceptually cleaner for "=" to do a simple
symbol definition with no implied evaluation, and we no longer need the clunky
"later" function, with its weird semantics, to defer evaluation.